# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkDeleteRequest": {
        "properties": {
          "patterns": {
            "items": {
              "type": "string"
            },
            "title": "Patterns",
            "type": "array"
          }
        },
        "required": [
          "patterns"
        ],
        "title": "BulkDeleteRequest",
        "type": "object"
      },
      "BulkDeleteResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "deleted": {
            "items": {
              "type": "string"
            },
            "title": "Deleted",
            "type": "array"
          },
          "not_found": {
            "items": {
              "type": "string"
            },
            "title": "Not Found",
            "type": "array"
          }
        },
        "required": [
          "deleted",
          "not_found",
          "count"
        ],
        "title": "BulkDeleteResponse",
        "type": "object"
      },
      "BulkImportRequest": {
        "properties": {
          "shortcuts": {
            "items": {
              "$ref": "#/components/schemas/ShortcutCreate"
            },
            "title": "Shortcuts",
            "type": "array"
          }
        },
        "required": [
          "shortcuts"
        ],
        "title": "BulkImportRequest",
        "type": "object"
      },
      "BulkImportResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "inserted": {
            "title": "Inserted",
            "type": "integer"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "title": "Skipped",
            "type": "array"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "inserted",
          "updated",
          "skipped",
          "count"
        ],
        "title": "BulkImportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShortcutCreate": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ShortcutType",
            "default": "static"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "default": "public"
          }
        },
        "required": [
          "pattern",
          "target"
        ],
        "title": "ShortcutCreate",
        "type": "object"
      },
      "ShortcutListMeta": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total"
        ],
        "title": "ShortcutListMeta",
        "type": "object"
      },
      "ShortcutListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ShortcutRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ShortcutListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ShortcutListResponse",
        "type": "object"
      },
      "ShortcutRead": {
        "properties": {
          "access_count": {
            "default": 0,
            "title": "Access Count",
            "type": "integer"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ShortcutType"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility",
            "default": "public"
          }
        },
        "required": [
          "pattern",
          "type",
          "target"
        ],
        "title": "ShortcutRead",
        "type": "object"
      },
      "ShortcutType": {
        "enum": [
          "static",
          "dynamic",
          "user-dynamic"
        ],
        "title": "ShortcutType",
        "type": "string"
      },
      "ShortcutUpdate": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortcutType"
              },
              {
                "type": "null"
              }
            ]
          },
          "visibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Visibility"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ShortcutUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Visibility": {
        "enum": [
          "public",
          "unlisted",
          "private",
          "team"
        ],
        "title": "Visibility",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Async FastAPI backend for Redirector (EPIC-01). Shortcuts, upstreams, resolve hot path, auth, metrics, QR.",
    "title": "Redirector API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/shortcuts": {
      "get": {
        "operationId": "list_shortcuts_api_v1_shortcuts_get",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "description": "Search term across pattern and target",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Search term across pattern and target",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Filter by tag",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": "",
              "description": "Filter by tag",
              "title": "Tag",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated_at",
              "pattern": "^(updated_at|created_at|popular)$",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortcutListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List shortcuts",
        "tags": [
          "shortcuts"
        ]
      },
      "post": {
        "operationId": "create_shortcut_api_v1_shortcuts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortcutCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortcutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create shortcut",
        "tags": [
          "shortcuts"
        ]
      }
    },
    "/api/v1/shortcuts/bulk-delete": {
      "post": {
        "operationId": "bulk_delete_shortcuts_api_v1_shortcuts_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk delete shortcuts",
        "tags": [
          "shortcuts"
        ]
      }
    },
    "/api/v1/shortcuts/bulk-import": {
      "post": {
        "description": "Upsert shortcuts by pattern (insert new, overwrite existing). Invalid\nrows are skipped, never fatal \u2014 the response reports every outcome.",
        "operationId": "bulk_import_shortcuts_api_v1_shortcuts_bulk_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk import shortcuts",
        "tags": [
          "shortcuts"
        ]
      }
    },
    "/api/v1/shortcuts/{pattern}": {
      "delete": {
        "operationId": "delete_shortcut_api_v1_shortcuts__pattern__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pattern",
            "required": true,
            "schema": {
              "title": "Pattern",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete shortcut",
        "tags": [
          "shortcuts"
        ]
      },
      "get": {
        "operationId": "get_shortcut_api_v1_shortcuts__pattern__get",
        "parameters": [
          {
            "in": "path",
            "name": "pattern",
            "required": true,
            "schema": {
              "title": "Pattern",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortcutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get shortcut details",
        "tags": [
          "shortcuts"
        ]
      },
      "patch": {
        "operationId": "update_shortcut_api_v1_shortcuts__pattern__patch",
        "parameters": [
          {
            "in": "path",
            "name": "pattern",
            "required": true,
            "schema": {
              "title": "Pattern",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortcutUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortcutRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update shortcut",
        "tags": [
          "shortcuts"
        ]
      }
    },
...
  },
...
}